Add visual representation of inventory in tablet
Inventory UI instantiates icons based on contents of inventory.
Whole list is cleared and redone when anything at all changes, might
be a good idea to optimize that a bit later.

When camera is transitioning to and from tabled, set GameState to
locked where no inputs work.

Various small edits and fixes.

+++ VirtualWorld/Assets/Scripts/Character/InventoryItem.cs
@@ -7,9 +7,9 @@ namespace Items
     public class InventoryItem
     {
         public Item item;
-        public double amount;
+        public int amount;
 
-        public InventoryItem(Item item, double amount)
+        public InventoryItem(Item item, int amount)
         {
             this.item = item;
             this.amount = amount;
